Feature/encrypt scrubber

+++ src/dcm/agent/cmd/scrubber.py
@@ -1,9 +1,21 @@
 import argparse
+import hashlib
 import os
 import pwd
+import random
 import re
+import string
+import subprocess
 import sys
 import tarfile
+import tempfile
+import uuid
+
+import Crypto.Cipher.AES as AES
+import Crypto.Random as Random
+import Crypto.PublicKey.RSA as RSA
+
+import dcm.agent.config as config
 
 
 opts_msg = """DCM Agent Image Preparation.
@@ -20,6 +32,11 @@ To backup all of the information that it will remove please use the
 it.  This file can then be untarred to restore the removed files.  However,
 this file should be manually copied off the server and then safely removed
 before the image creation occurs.
+
+It is recommended that this file is encrypted using a public key on this system.
+It can then be decrypted by using the matching private key which should be
+safely stored in a location off of this system.  To encrypt the recovery tarball
+use the -e option.
 """
 
 _g_tarfile_output_message = """
@@ -31,14 +48,27 @@ the server and then securely delete it!
 ******************************************************************************
 """
 
+_g_public_key_message = """
+******************************************************************************
+When creating a restoring tarfile it is recommended that this file be encrypted
+with a public key.  This way if it is burnt into a child VM image it cannot
+be seen by any parties that may boot that image in the future.  To restore the
+rescue file the associated private key (which should not be on this system can
+be used.
+******************************************************************************
+"""
+
 
 def setup_command_line_parser():
     parser = argparse.ArgumentParser(description=opts_msg)
     parser.add_argument("-v", "--verbose",
                         help="Increase the amount of output.",
                         action='count', default=1)
-    parser.add_argument("-r", "--rescue-tar",
+    parser.add_argument("-r", "--rescue-file",
                         help="Create a tarball that can be used to recover the secrets that this will erase.",
+                        default=None)
+    parser.add_argument("-e", "--public-key",
+                        help="A path to the public encryption key that will be used to encrypt this file.",
                         default=None)
     parser.add_argument("-c", "--cloud-init",
                         help="Delete cloud-init cache and logs.",
@@ -67,6 +97,9 @@ def setup_command_line_parser():
     parser.add_argument("-b", "--batch",
                         help="Run the program without interrupting the user.  This could cause their to be no rescue file.",
                         action="store_true")
+    parser.add_argument("-D", "--dry-run",
+                        help="Run the program without actually deleting any files.",
+                        action="store_true")
     return parser
 
 
@@ -91,6 +124,9 @@ def get_secure_delete():
         if tar is not None:
             tar.add(path)
 
+        if opts.dry_run:
+            return
+
         console_output(opts, 2, "Securely deleting %s" % path)
         rc = os.system("%s  -f -z %s" % (srm_path, path))
         if rc != 0:
@@ -100,6 +136,9 @@ def get_secure_delete():
         if tar is not None:
             tar.add(path)
         console_output(opts, 2, "Deleting %s" % path)
+
+        if opts.dry_run:
+            return
         os.remove(path)
 
     if srm_path:
@@ -190,11 +229,14 @@ def clean_agent_files(opts, tar):
                       '/tmp/error.log',
                       '/tmp/installer.sh']
 
+    conf = config.AgentConfig(config.get_config_files())
+    log_dir = os.path.join(conf.storage_base_dir, "logs")
+
     if not opts.agent_running:
-        files_to_clean.append('/dcm/logs/agent.log')
-        files_to_clean.append('/dcm/logs/agent.log.job_runner')
-        files_to_clean.append('/dcm/logs/agent.log.wire')
-        files_to_clean.append('/dcm/secure/agentdb.sql')
+        files_to_clean.append(os.path.join(log_dir, 'agent.log'))
+        files_to_clean.append(os.path.join(log_dir, 'agent.log.job_runner'))
+        files_to_clean.append(os.path.join(log_dir, 'agent.log.wire'))
+        files_to_clean.append(conf.storage_dbfile)
 
     for f in files_to_clean:
         if os.path.exists(f):
@@ -228,22 +270,194 @@ def clean_dhcp_leases(opts, tar):
                     secure_delete(opts, tar, file)
 
 
+def get_get_public_key_path(opts):
+    if opts.batch or opts.public_key is not None:
+        return opts.public_key
+
+    sys.stdout.write(_g_public_key_message)
+    sys.stdout.write("Would you like to encrypt with the public key (Y/n)? ")
+    sys.stdout.flush()
+    answer = sys.stdin.readline().strip()
+    if answer.lower() != 'y' and answer.lower() != "yes":
+        return None
+
+    key_path = os.path.expanduser("~/.ssh/id_rsa.pub")
+    sys.stdout.write(
+        "Please enter the path to the public key to use for encryption (%s): "
+        % key_path)
+    sys.stdout.flush()
+    answer = sys.stdin.readline().strip()
+    if answer:
+        key_path = answer
+
+    if not os.path.exists(key_path):
+        raise Exception("The key path %s does not exist." % key_path)
+
+    return key_path
+
+
+def get_public_key_data(opts):
+    if opts.rescue_file is None:
+        # nothing to encrypt if there is no tar
+        return None
+    public_key_path = get_get_public_key_path(opts)
+    if not public_key_path:
+        return None
+
+    console_output(opts, 1, "Using the public key %s" % public_key_path)
+    try:
+        with open(public_key_path, "r") as fptr:
+            return fptr.readline()
+    except IOError:
+        raise Exception("The public key file %s could not be read."
+                        % public_key_path)
+
+
+def get_rescue_path(opts):
+    if opts.rescue_file is not None or opts.batch:
+        return os.path.abspath(opts.rescue_file)
+    sys.stdout.write("Please enter the location of the rescue tarfile:")
+    sys.stdout.flush()
+    rescue_path = sys.stdin.readline().strip()
+    if not rescue_path:
+        return None
+    rescue_path = os.path.abspath(rescue_path)
+    return rescue_path
+
+
+def get_tar(opts, rescue_path):
+    if rescue_path is None:
+        return None, None
+    console_output(opts, 1, "Using the rescue file %s" % rescue_path)
+    osf, tarfile_path = tempfile.mkstemp()
+    os.close(osf)
+    tar = tarfile.open(tarfile_path, "w:gz")
+    return tarfile_path, tar
+
+
+def generate_symmetric_key():
+    symmetric_key = str(uuid.uuid4()) + ''.join(random.choice(
+        string.ascii_letters + string.digits + "-_!@#^(),.=+")
+        for _ in range(10))
+    return symmetric_key
+
+
+def derive_key_and_iv(password, salt, key_length, iv_length):
+    d = d_i = b''
+    while len(d) < key_length + iv_length:
+        d_i = hashlib.md5(d_i + password + salt).digest()
+        d += d_i
+    return d[:key_length], d[key_length:key_length+iv_length]
+
+
+def encrypt(in_file, out_file, password, key_length=32):
+    bs = AES.block_size
+    salted_bytes = 'Salted__'.encode()
+    salt = Random.new().read(bs - len(salted_bytes))
+    key, iv = derive_key_and_iv(password.encode(), salt, key_length, bs)
+    cipher = AES.new(key, AES.MODE_CBC, iv)
+    out_file.write(salted_bytes + salt)
+    finished = False
+    while not finished:
+        chunk = in_file.read(1024 * bs)
+        if len(chunk) == 0 or len(chunk) % bs != 0:
+            padding_length = (bs - len(chunk) % bs) or bs
+            chunk += padding_length * chr(padding_length).encode()
+            finished = True
+        out_file.write(cipher.encrypt(chunk))
+
+
+def _write_temp_file(data):
+    osf, temp_file_path = tempfile.mkstemp()
+    try:
+        os.write(osf, data)
+        return temp_file_path
+    finally:
+        os.close(osf)
+
+
+def encrypt_symmetric_key_with_public_key(symmetric_key, public_key):
+    rsa_pk = RSA.importKey(public_key)
+    pk = rsa_pk.publickey()
+    pem_pub = pk.exportKey(format='PEM')
+    public_key_file = _write_temp_file(pem_pub)
+
+    try:
+        openssl_binary_location = "openssl"
+        args = [openssl_binary_location,
+                'rsautl', '-encrypt', '-pubin',
+                '-inkey', public_key_file]
+        print(' '.join(args))
+        process = subprocess.Popen(' '.join(args),
+                                   stdin=subprocess.PIPE,
+                                   stdout=subprocess.PIPE,
+                                   stderr=subprocess.PIPE,
+                                   shell=True)
+        (stdout, stderr) = process.communicate(symmetric_key.encode())
+        rc = process.wait()
+        if rc != 0:
+            raise Exception("Public key encryption failed: %s" % stderr)
+        return stdout
+    finally:
+        os.remove(public_key_file)
+
+
+def encrypt_with_key(tarfile_path, public_key):
+    if public_key is None or tarfile_path is None:
+        return tarfile_path, None
+
+    # first generate the symmetric key to encrypt
+    symmetric_key = generate_symmetric_key()
+    encrypted_key = encrypt_symmetric_key_with_public_key(
+        symmetric_key, public_key)
+
+    osf, temp_path = tempfile.mkstemp()
+    try:
+        with open(tarfile_path, "rb") as tar_fptr,\
+                os.fdopen(osf, "wb") as out_fptr:
+            encrypt(tar_fptr, out_fptr, symmetric_key)
+        return temp_path, encrypted_key
+    finally:
+        os.remove(tarfile_path)
+
+
+def make_rescue_file(data_tar_path, rescue_file_destination_path,
+                     encrypted_key=None, public_key=None):
+
+    # find the recovery
+    recovery_script_path = os.path.join(config.get_python_script_dir(),
+                                        "recovery.sh")
+    temp_key_path = None
+    temp_key_name_path = None
+    try:
+        tar = tarfile.open(rescue_file_destination_path, "w:gz")
+
+        if encrypted_key is not None:
+            temp_key_path = _write_temp_file(encrypted_key)
+            tar.add(temp_key_path, arcname='key')
+
+        if public_key is not None:
+            temp_key_name_path = _write_temp_file(public_key.encode())
+            tar.add(temp_key_name_path, arcname='public_key')
+
+        tar.add(data_tar_path, arcname='data.enc')
+        tar.add(recovery_script_path, arcname='recovery.sh')
+        tar.close()
+    finally:
+        if temp_key_path is not None:
+            os.remove(temp_key_path)
+        if temp_key_name_path is not None:
+            os.remove(temp_key_name_path)
+        os.remove(data_tar_path)
+
+
 def main(args=sys.argv):
     parser = setup_command_line_parser()
     opts = parser.parse_args(args=args[1:])
 
-    tarfile_path = opts.rescue_tar
-    if tarfile_path is None and not opts.batch:
-        sys.stdout.write("Please enter the location of the rescue tarfile:")
-        sys.stdout.flush()
-        tarfile_path = sys.stdin.readline().strip()
-
-    if tarfile_path is not None:
-        tarfile_path = os.path.abspath(tarfile_path)
-        console_output(opts, 1, "Using the rescue file %s" % tarfile_path)
-        tar = tarfile.open(tarfile_path, "w:gz")
-    else:
-        tar = None
+    public_key_data = get_public_key_data(opts)
+    rescue_path = get_rescue_path(opts)
+    (tarfile_path, tar) = get_tar(opts, rescue_path)
 
     try:
         if opts.history:
@@ -259,12 +473,15 @@ def main(args=sys.argv):
         if opts.agent:
             clean_agent_files(opts, tar)
         if opts.agent_token:
-            secure_delete(opts, tar, "/dcm/secure/token")
+            try:
+                secure_delete(opts, tar, "/dcm/secure/token")
+            except FileNotFoundError:
+                console_output(opts, 1, "The token file does not exist.")
         general_cleanup(opts, tar)
     except BaseException as ex:
         if tar is not None:
             tar.close()
-            os.remove(opts.rescue_tar)
+            os.remove(tarfile_path)
         console_output(opts, 0, "Error: " + str(ex))
         if opts.verbose > 1:
             raise
@@ -272,6 +489,13 @@ def main(args=sys.argv):
     else:
         if tar is not None:
             tar.close()
+            tarfile_path, encrypted_key =\
+                encrypt_with_key(tarfile_path, public_key_data)
+
+            make_rescue_file(tarfile_path, rescue_path,
+                          encrypted_key=encrypted_key,
+                          public_key=public_key_data)
+
             console_output(
                 opts, 0,
-                _g_tarfile_output_message % tarfile_path)
+                _g_tarfile_output_message % rescue_path)
